Web: Reorder Introduction contents
Menu structure matches the structure of the boxes.

There are some things I would like to do but don't dare without
prior discussion (in whatever circle and form):

TODO: I would really like to move Essay to Introduction and
replace the Background menu intem with that.
It has to be discussed if it can be moved away from Manuals
(it can still be linked from there but would be purged from the menu

TODO: I would like to Move Productions and Reviews to Community.