Issue 3845: Web: Rewrite Closing of "Text input"
Discussion indicated that "Text input" - while very good in
itself - isn't sufficient to let all readers get the idea
of "editing and compiling".
Therefore I added a new page discussing that concept.

This commit rewrites the closing of "Text input" to make
that clear and @ref to the new page.