Factor out Beaming_pattern::fix_extremal_stems
We can't process the extremal stems from within the loop
because some operations look left and right of the current stem
(which would cause access violations in the vectors).

Factoring out the post-processing of the extremal stems is not
necessary but it makes the main beamify function more readable
and maintainable.